Added FINAL_DUMMY state for general-purpose NFA/DFA use. StNonFinal.NOT_FINAL migrated to StateTag. FINAL_DUMMY and NOT_FINAL are now constants in StateTag interface. These are special tags that DO NOT produce any domain!
